requirements.txt
# Core dependencies - using specific versions with Python 3.12 wheels
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4
scikit-learn==1.3.2
pytz==2023.3.post1
python-dateutil==2.8.2

# Technical analysis - using the latest version
pandas-ta==0.3.14b

# API and Web
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.4.2
requests==2.31.0
aiohttp==3.9.0

# Zerodha integration
kiteconnect==4.1.0

# Concurrency and async
asyncio==3.4.3
aiofiles==23.2.1

# Logging and monitoring
loguru==0.7.2

# Data formats
openpyxl==3.1.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.0

# Deployment
python-dotenv==1.0.0
gunicorn==21.2.0